+++ src/data/shortlist.h
@@ -43,8 +43,8 @@ public:
   Shortlist(const std::vector<WordIndex>& indices);
   virtual ~Shortlist();
   
-  virtual WordIndex reverseMap(size_t batchIdx, size_t beamIdx, int idx) const;
-  virtual WordIndex tryForwardMap(size_t batchIdx, size_t beamIdx, WordIndex wIdx) const;
+  virtual WordIndex reverseMap(int batchIdx, int beamIdx, int idx) const;
+  virtual WordIndex tryForwardMap(int batchIdx, int beamIdx, WordIndex wIdx) const;
 
   virtual void filter(Expr input, Expr weights, bool isLegacyUntransposedW, Expr b, Expr lemmaEt);
   virtual Expr getIndicesExpr(int batchSize, int currBeamSize) const;
@@ -76,8 +76,8 @@ private:
 
 public:
   LSHShortlist(int k, int nbits);
-  virtual WordIndex reverseMap(size_t batchIdx, size_t beamIdx, int idx) const override;
-  virtual WordIndex tryForwardMap(size_t batchIdx, size_t beamIdx, WordIndex wIdx) const override;
+  virtual WordIndex reverseMap(int batchIdx, int beamIdx, int idx) const override;
+  virtual WordIndex tryForwardMap(int batchIdx, int beamIdx, WordIndex wIdx) const override;
 
   virtual void filter(Expr input, Expr weights, bool isLegacyUntransposedW, Expr b, Expr lemmaEt) override;
   virtual Expr getIndicesExpr(int batchSize,int currBeamSize) const override;
